requirements.txt
playwright==1.51.0
gspread==6.1.4
google-auth==2.38.0
google-api-python-client==2.164.0
Jinja2==3.1.6
python-dotenv==1.1.0
